claude-mcp/mcp_server.py
import uvicorn
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel
from typing import Dict, List, Optional
import json
import requests
from dotenv import load_dotenv
import os

load_dotenv()

app = FastAPI()

class FunctionCall(BaseModel):
    name: str
    parameters: Dict

class MCPRequest(BaseModel):
    function_calls: List[FunctionCall]

@app.post("/rpc")
async def handle_rpc(request: MCPRequest):
    results = []
    
    for call in request.function_calls:
        try:
            if call.name == 'brave_web_search':
                # Implement web search functionality
                results.append({
                    "status": "success",
                    "result": f"Search results for: {call.parameters['query']}"
                })
                
            elif call.name == 'create_repository':
                # Implement GitHub repository creation
                results.append({
                    "status": "success",
                    "result": f"Created repository: {call.parameters['name']}"
                })
                
            else:
                results.append({
                    "status": "error",
                    "error": f"Unknown function: {call.name}"
                })
                
        except Exception as e:
            results.append({
                "status": "error",
                "error": str(e)
            })
    
    return {"results": results}

@app.get("/schema")
async def get_schema():
    return {
        "functions": [
            {
                "name": "brave_web_search",
                "description": "Performs a web search",
                "parameters": {
                    "type": "object",
                    "properties": {
                        "query": {
                            "type": "string",
                            "description": "Search query"
                        }
                    },
                    "required": ["query"]
                }
            },
            {
                "name": "create_repository",
                "description": "Creates a GitHub repository",
                "parameters": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string",
                            "description": "Repository name"
                        }
                    },
                    "required": ["name"]
                }
            }
        ]
    }

def main():
    uvicorn.run(app, host="0.0.0.0", port=8000)

if __name__ == "__main__":
    main()